py/requirements.txt
zxcvbn >= 4.4.28
tqdm >= 4.45.0
stopwatch.py >= 1.0.1
